+++ src/main/scala/xitrum/view/Responder.scala
@@ -191,7 +191,7 @@ trait Responder extends Js with Flash with GetActionClassDefaultsToCurrentAction
    * It makes debugging a pain.
    */
   def respondJson(ref: AnyRef): ChannelFuture = {
-    val json = Json.generate(ref)
+    val json = Json.serialize(ref)
     respondText(json, "application/json")
   }
 
@@ -203,7 +203,7 @@ trait Responder extends Js with Flash with GetActionClassDefaultsToCurrentAction
    * Content-Type header is set to "application/javascript".
    */
   def respondJsonP(ref: AnyRef, function: String): ChannelFuture = {
-    val json = Json.generate(ref)
+    val json = Json.serialize(ref)
     val text = function + "(" + json + ");\r\n"
     respondJs(text)
   }